Modificacion de BottomNavigation, se cambio el fragment del registro por un activity
