+++ app/src/main/res/menu/home_menu.xml
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
-        android:id="@+id/homemenu_home"
+        android:id="@+id/homemenu_search"
         android:enabled="true"
         android:icon="@drawable/insect"
-        android:title="@string/MisReg"
+        android:title="@string/search"
         />
     <item
-        android:id="@+id/homemenu_registros"
+        android:id="@+id/homemenu_home"
         android:enabled="true"
         android:icon="@drawable/home"
         android:title="@string/home"
